Fix last remaining `emulate -L zsh -[options]`
Change the `emulate -L zsh -eu` at the top of the definition of the
function `path-lookup` to `emulate -L zsh; set -eu`.

This was the last remaining `emulate -L zsh -[options]` line, of several
that I had previously so changed.

The reason for changing these lines is that, apparently, including
shell command-line options after the shell name causes the `-L` option
to `emulate` to be ignored, resulting in the shell options being reset
globally.