Add `with-no-rprompt` function
Add a `with-no-rprompt` shell function, which runs a given command in an
environment where the shell parameter `ZSHRC_NO_RPROMPT` is set to `y`.

The motivating use-case for this function is the invocation
`with-no-rprompt zsh`, which opens a new zsh session that will not use a
`RPROMPT` (a shell prompt on the right-hand side of the display), at
least unless subsequently instructed otherwise.