Integrate `TIMEFMT` into dynamic prompts system
In the `set-dynamic-prompts` function, in addition to setting the
special shell parameters that control the formats of the shell prompts,
now also set the special shell parameter `TIMEFMT`, which controls the
format of the running-time reports zsh emits for commands that (a) are
prefixed with `time` or (b) occupy more than `$REPORTTIME` seconds of
CPU time.

This allows the format of these reports to be controlled with a
`time-report-fmt` entry in the `~/.zsh/prompt-style` configuration file,
which is local to each user-account.

Because, in the current format of the `prompt-style` file, the values of
entries cannot contain whitespace, the following translations will be
performed on the value of a `time-report-fmt` entry in the
`prompt-style` file:

  - Low-lines not enclosed in parentheses (`_`) will be translated to
    spaces.

  - Low-lines enclosed in parentheses (`(_)`) will be translated to
    horizontal tabs.

Furthermore, because UTF-8–encoded non–US-ASCII Unicode code-points are
not presently supported in `TIMEFMT` (see also the function
`have-zsh-supporting-Unicode-TIMEFMT`), the following translations will
also be performed on the value of a `time-report-fmt` entry in the
`prompt-style` file unless UTF-8–encoded non–US-ASCII Unicode
code-points appear to be supported in `TIMEFMT`:

  - Typographical single quotation marks (`‘`, `’`) will be translated
    to grave accents (`` ` ``).

  - Typographical double quotation marks (`“`, `”`) will be translated
    to straight double quotation marks (`"`).

  - Em-dashes (`—`) will be translated to double hyphen-minuses (`--`).