Add command running-time reporting system
Add a command running-time reporting system, which extends the
functionality of `REPORTTIME` to optionally also report timing
information if a command runs for at least `$REPORTTIME` seconds in
total, rather than only if a command occupies at least `$REPORTTIME`
seconds of CPU time.

The command running-time reporting system modifies the behavior of the
special shell parameter `REPORTTIME`, depending on the value of the
shell parameter `REPORTTIME_TYPE`, which may be `cpu-time`, `run-time`,
or `both`.

When `$REPORTTIME_TYPE` is `cpu-time`, the behavior of `REPORTTIME` will
be left at its default, which is to report running-time information for
any command that occupies at least `$REPORTTIME` seconds of CPU time.

When `$REPORTTIME_TYPE` is `run-time`, for any command that runs for at
least `$REPORTTIME` seconds, regardless of how much time the command
spent actively consuming CPU power, the total running-time of that
command will be reported.

When `$REPORTTIME_TYPE` is `both`, both the `cpu-time` and `run-time`
behaviors will be enabled. This is the default.