Add `run-secure-base` function
Add a zsh function `run-secure-base`, which combines the
`select-secure-executable` and `run-by-unambiguous-basename` functions
to run a program by unambiguous basename, ensuring that it is "secure"
as determined by `file-is-secure`, and emitting warning messages if
matching insecure executables are found in `$PATH`.

This change was found uncommitted in a backup of this repository and
an associated working tree, with this `zshrc` file dated 2014-12-05 by
modification time.