Cull nonexistent dirs in module_path, fpath sooner
Filter nonexistent directories out of the values of the shell
parameters `module_path` and `fpath` before using those values in
checking whether the directories to which they refer are "secure" as
determined by the `assert-file-is-secure` function.

This was already done in the subsequent "Shell parameters" section,
but doing this here prevents the aforementioned checks from erroring
in the case that the aforementioned path array shell parameters
include nonexistent directories.