Ensure compl fns run without POSIX_IDENTIFIERS
After initializing the completion system by calling the `compinit` zsh
function, append the string "NO_POSIX_IDENTIFIERS" to the
`_comp_options` shell array parameter, instructing the completion
system to disable the shell option `POSIX_IDENTIFIERS` for individual
completion functions.

The completion functions, in several places, use the syntax
`$#parameter` to get the length of a shell parameter, which syntax is
an abbreviated form of `${#parameter}` that is disabled when the
shell option `POSIX_IDENTIFIERS` is enabled.