Report variance from recorded, not reported, time
Change the initialization timing system to report certain execution
duration measurements if the variance between total and recorded,
rather than total and reported, execution durations equals or exceeds
a given threshold value.

This eliminates some noise, to wit, the report of variance between
total and reported time, which was largely redundant with the report
of total time.