cmd-running-time system: Ignore time between cmds
In the command running-time reporting system, rather than setting the
system into the ready state once, just before the prompt is first
printed, and leaving it in the ready state, instead set it into the
ready state just before a command is run and take it out of the ready
state after the command finishes running, just after it reports the
command's running time (if it does so).

This prevents the system from tracking and reporting time between
command executions, i.e., time when the shell is idle or when the user
is writing a command.

This issue seems to have appeared in my hiatus from zsh; it may have
been caused by a change between zsh versions.