fn view-zshfn: Fix misuse of `readonly`
In the zsh function `view-zshfn`, I had declared a shell parameter `f`
as `readonly` inside a `for` loop, under the apparent mistaken belief
that the parameter would be block-scoped, when it is in actuality
function-scoped.

Because the parameter lived through each loop iteration, it would be
reassigned to on each loop iteration after the first, raising an
error.

My suspicion that the version of zsh I wrote this in didn't enforce
`readonly` grows.