Prompt: Abbreviate paths to fit cfg'ble width
Provide a facility for configuring the dynamic prompts subsystem to
abbreviate leading components of paths in the right-hand-side prompt
such as to try to fit that prompt within a configurable portion of the
terminal window.

This functionality defaults to being disabled. It can be enabled and
configured by including in the `prompt-style` file the line —

    rprompt-abbrev-threshold-term-width-divisor <divisor>

— which configures this functionality to abbreviate path segments
inside right-hand-side prompts such as to try to fit those prompts
within (( COLUMNS / <divisor> )) columns.