Delete obsolete `git-hub` set-up code
Delete a block of code that was used to initialize the shell
environment for the use of Ingy döt Net's `git-hub` program. This code
is obsolete in multiple ways: I now install `git-hub` system-wide; I
no longer keep my checkout of `git-hub`'s Git repository at the path
where this code expects it to be located; and `git-hub` no longer
supports the shell environment initialization method used in this
code.