fn run-by-unambiguous-basename: don't purge PATH
In the zsh function `run-by-unambiguous-basename`, rather than
ensuring that the command to be run runs in an environment where the
directory wherein its executable resides is the only directory in the
`PATH`, instead ensure that said directory is the first directory in
the `PATH`, with all other directories normally in the `PATH` being
allowed to remain in the `PATH`.

This fixes an issue whereby programs run by
`run-by-unambiguous-basename` would be unable to run other programs
whose executables resided in different directories.