Add fn `e` for opening a text editor
In the `zshrc` script, add a zsh function `e` which opens a text
editor considered to be the user's desired default (except with any
`-nofork` suffix stripped), passing along all arguments.

The desired default text editor is considered to be that specified by
the environment variable `VISUAL`, or by the environment variable
`EDITOR` if `VISUAL` is rejected by `select-secure-cmd`.